Adicionado partes do CSS

+++ Conversor.js
@@ -7,18 +7,18 @@ var fObj = document.convert.ftemp, cObj = document.convert.ctemp, kObj = documen
 if (direction == "ftoc") 
 {
     cObj.value = Math.round((fObj.value - 32) * (5/9));
-	kObj.value = Math.round((parseInt(cObj.value) + 459.67) * (5/9));
+	kObj.value = Math.round((parseFloat(cObj.value) + 459.67) * (5/9));
 } 
 else if (direction == "ktof") 
  {
-	fObj.value = Math.round((parseInt(cObj.value) * (9/5)) -  459.67);
+	fObj.value = Math.round((parseFloat(cObj.value) * (9/5)) -  459.67);
 	cObj.value = Math.round((fObj.value - 32) * (5/9));
 	
  }
 else
  {
-	fObj.value = Math.round((parseInt(cObj.value) * (9/5)) + 32);
-	kObj.value = Math.round((parseInt(cObj.value) + 273));
+	fObj.value = Math.round((parseFloat(cObj.value) * (9/5)) + 32);
+	kObj.value = Math.round((parseFloat(cObj.value) + 273));
  }
 }
